requirements.txt
python-dotenv~=0.14.0
mysql-connector-python~=8.0.21
pytelegrambotapi